Move common processing to PartsMain
