+++ styles/PartsMain.module.scss
